+++ src/engine/shared/json.h
@@ -5,9 +5,9 @@
 
 const struct _json_value *json_object_get(const json_value *object, const char *index);
 const struct _json_value *json_array_get(const json_value *array, int index);
-int json_array_length(const json_value *array);
+unsigned int json_array_length(const json_value *array);
 const char *json_string_get(const json_value *string);
-int json_int_get(const json_value *integer);
+int64_t json_int_get(const json_value *integer);
 int json_boolean_get(const json_value *boolean);
 
 char *EscapeJson(char *pBuffer, int BufferSize, const char *pString);
